requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
vertexai==1.*
google-generativeai==0.8.4
